+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/DummyOtherProvider.groovy
@@ -1,7 +1,6 @@
 package ru.vyarus.dropwizard.guice.support.feature
 
 import javax.ws.rs.ext.ContextResolver
-import javax.ws.rs.ext.Provider
 
 /**
  * Check other provider type installation.
@@ -9,8 +8,7 @@ import javax.ws.rs.ext.Provider
  * @author Vyacheslav Rusakov 
  * @since 14.10.2014
  */
-@Provider
-class DummyOtherProvider implements ContextResolver{
+class DummyOtherProvider implements ContextResolver {
 
     @Override
     Object getContext(Class type) {